@@ -2,7 +2,10 @@
 
 from __future__ import annotations
 import functools
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
 
 
 def report_bad_line(line_parser: Callable) -> Callable:
