@@ -9,12 +9,13 @@ See:
 from __future__ import annotations
 import sys
 from functools import wraps
-from typing import TYPE_CHECKING, Callable, Optional, Union
+from typing import TYPE_CHECKING, Optional, Union
 
 import numpy as np
 from scipy import stats
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from numpy import float64, ndarray
 
 
