@@ -56,7 +56,7 @@ def zip_repeater(
             yield it, rpt
     else:
         i = -1
-        for i, (it, rpt) in enumerate(zip(iterable, repeatable)):
+        for i, (it, rpt) in enumerate(zip(iterable, repeatable, strict=False)):
             yield it, rpt
         # Require lengths to match
         if i + 1 != rpt_len:
