@@ -8,9 +8,10 @@ import numpy as np
 import pandas as pd
 
 from .descriptives import weighted_median
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from cnvlib.cnary import CopyNumArray
     from pandas.core.frame import DataFrame
     from pandas.core.series import Series
