@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import logging
 from collections import OrderedDict
-from typing import Any, Callable, Optional, Union, TYPE_CHECKING
+from typing import Any, Optional, Union, TYPE_CHECKING
 
 import numpy as np
 import pandas as pd
@@ -17,6 +17,7 @@ from .subtract import subtract
 from .subdivide import subdivide
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from cnvlib.cnary import CopyNumArray
     from cnvlib.vary import VariantArray
     from numpy import ndarray
@@ -43,7 +44,7 @@ class GenomicArray:
         # Validation
         if (
             data_table is None
-            or (isinstance(data_table, (list, tuple)) and not len(data_table))
+            or (isinstance(data_table, list | tuple) and not len(data_table))
             or (isinstance(data_table, pd.DataFrame) and not len(data_table.columns))
         ):
             data_table = self._make_blank()
@@ -71,7 +72,7 @@ class GenomicArray:
 
             recast_cols = {
                 col: dtype
-                for col, dtype in zip(self._required_columns, self._required_dtypes)
+                for col, dtype in zip(self._required_columns, self._required_dtypes, strict=False)
                 if not ok_dtype(col, dtype)
             }
             if recast_cols:
@@ -83,7 +84,7 @@ class GenomicArray:
     @classmethod
     def _make_blank(cls) -> pd.DataFrame:
         """Create an empty dataframe with the columns required by this class."""
-        spec = list(zip(cls._required_columns, cls._required_dtypes))
+        spec = list(zip(cls._required_columns, cls._required_dtypes, strict=False))
         try:
             arr = np.zeros(0, dtype=spec)
             return pd.DataFrame(arr)
