@@ -65,7 +65,7 @@ def unpack_range(a_range: Union[str, Sequence]) -> Region:
         if ":" in a_range and "-" in a_range:
             return from_label(a_range, keep_gene=False)  # type: ignore
         return Region(a_range, None, None)
-    if isinstance(a_range, (list, tuple)):
+    if isinstance(a_range, list | tuple):
         if len(a_range) == 3:
             return Region(*a_range)
         if len(a_range) == 4:
